Changes to sync and layout
