+++ move/lib/screens/activityPage.dart
@@ -187,136 +187,6 @@ class _ActivityPageState extends State<ActivityPage>
     }
   }
 
-
-  /*Future<void> _listCSVFiles() async {
-    Directory? downloadsDirectory ;
-    if (Platform.isAndroid) {
-      //downloadsDirectory = Directory('/storage/emulated/0/Download');
-      downloadsDirectory = await getApplicationDocumentsDirectory();
-    } else if (Platform.isIOS) {
-      downloadsDirectory = await getApplicationDocumentsDirectory();
-    }
-    if (downloadsDirectory != null) {
-      String downloadsPath = downloadsDirectory.path;
-      Directory downloadsDir = Directory(downloadsPath);
-      if (downloadsDir.existsSync()) {
-        List<FileSystemEntity> files = downloadsDir.listSync();
-
-        List<File> csvFiles = files
-            .where((file) => file is File && file.path.endsWith('.csv'))
-            .map((file) => file as File)
-            .where((file) => p.basename(file.path).startsWith("activity_")) // Filter by prefix
-            .toList();
-
-        List<String> fileNames = csvFiles.map((file) => p.basename(file.path)).toList();
-
-        for (File file in csvFiles) {
-          String timestamp = await _getSecondLineTimestamp(file);
-          String timestamp1 = timestamp.split(",")[0];
-          int timestamp2 = int.parse(timestamp1);
-          int updatedTimestamp = timestamp2*1000;
-          String fileName1 = p.basename(file.path);
-
-          DateTime timestampDateTime = DateTime.fromMillisecondsSinceEpoch(updatedTimestamp);
-          DateTime now = DateTime.now();
-          // print("......"+now.toString());
-          if(_tabController.index == 0){
-            String todayStr = _formatDate(now);
-            if (_formatDate(timestampDateTime) == todayStr) {
-              getFileData(fileName1);
-            }else{
-            }
-          }else if(_tabController.index == 1){
-            DateTime weekStart = now.subtract(Duration(days: 7));
-            if (timestampDateTime.isAfter(weekStart) && timestampDateTime.isBefore(now)) {
-              getFileData(fileName1); // Process the file data
-            }
-          }else if(_tabController.index == 2){
-            DateTime monthStart = now.subtract(Duration(days: 30));
-            if (timestampDateTime.isAfter(monthStart) && timestampDateTime.isBefore(now)) {
-              getFileData(fileName1); // Process the file data
-            }
-          }
-
-        }
-
-      }
-    }
-  }
-
-
-  Future<void> getFileData(String fileName) async {
-    Directory? downloadsDirectory;
-    String myData = '';
-
-    if (Platform.isAndroid) {
-      //downloadsDirectory = Directory('/storage/emulated/0/Download');
-      downloadsDirectory = await getApplicationDocumentsDirectory();
-
-    } else if (Platform.isIOS) {
-      downloadsDirectory = await getApplicationDocumentsDirectory();
-    }
-
-    if (downloadsDirectory != null) {
-      String filePath = '${downloadsDirectory.path}/$fileName'; // Replace 'your_file.csv' with your actual file name
-      File csvFile = File(filePath);
-
-
-      if (await csvFile.exists()) {
-        String fileContent = await csvFile.readAsString();
-        setState(() {
-          myData = fileContent;
-        });
-      }
-    }
-
-    //String myData = await rootBundle.loadString("assets/Temp_data.csv");
-    List<String> result = myData.split('\n');
-    //print(result);
-    timestamp = result.map((f) => f.split(",")[0]).toList();
-
-    avgActivity = result.map((f) => f.split(",")[1]).toList();
-
-    for(int i = 1; i< timestamp.length; i++){
-      int tempTimeStamp = 0;
-      int tempTimeStamp1 = 0;
-      int tempAvgActivity = 0;
-
-      tempTimeStamp = int.parse(timestamp[i]);
-      tempTimeStamp1 = tempTimeStamp*1000;
-      tempAvgActivity = int.parse(avgActivity[i]);
-
-      //DateTime getUTCTime = DateTime.fromMillisecondsSinceEpoch(tempTimeStamp1).toUtc();
-      var getUTCTime = DateTime.fromMillisecondsSinceEpoch(tempTimeStamp1).toUtc();
-      // Format the DateTime to remove the 'Z' and make it human-readable
-      String formattedDate = DateFormat("yyyy-MM-dd HH:mm:ss").format(getUTCTime);
-      // Parse the formatted date string back into a DateTime object
-      DateTime formattedDateTime = DateTime.parse(formattedDate);
-
-
-      setState((){
-        //print(DateTime.fromMillisecondsSinceEpoch(ActivityTimeStamp1).toString());
-        ActivityTrendsData.add(ActivityTrends(formattedDateTime, tempAvgActivity-1));
-        totalCount = totalCount + tempAvgActivity;
-        print("......"+timestamp.length.toString());
-        if( i == timestamp.length-1){
-          lastUpdatedTime = formattedDateTime;
-          Count = tempAvgActivity;
-        }
-      });
-    }
-    _saveValue();
-  }
-
-  // Save a value
-  _saveValue() async {
-    String lastDateTime = DateFormat('yyyy-MM-dd HH:mm:ss').format(lastUpdatedTime);
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    await prefs.setString('latestActivity', Count.toString());
-    await prefs.setString('lastUpdatedActivity', lastDateTime.toString());
-  }*/
-
-
   Future<void> _listCSVFiles() async {
     Directory? downloadsDirectory;
     if (Platform.isAndroid) {
